A bit of cleanup on start.sh and main.js

+++ app/start.sh
@@ -10,6 +10,15 @@ build_server() {
     go build -o anubot-server anubot/cmd/api-server
 }
 
+build_lib() {
+    if [ -e lib ]; then
+        echo removing lib
+        rm -r lib
+    fi
+    babel --presets es2015,react --out-dir lib src
+    node-sass --output lib/styles src/styles
+}
+
 watch_app_source() {
     babel --presets es2015,react --watch --out-dir lib src &
     BABEL_PID=$!
@@ -23,18 +32,26 @@ watch_app_styles() {
 }
 
 kill_watchers() {
-    echo tearing down babel pid $BABEL_PID
-    kill $BABEL_PID
-    echo tearing down node-scss pid $NODE_PID
-    kill $NODE_PID
+    if [ -n "$BABEL_PID" ]; then
+        echo tearing down babel pid $BABEL_PID
+        kill $BABEL_PID
+    fi
+    if [ -n "$NODE_PID" ]; then
+        echo tearing down node-scss pid $NODE_PID
+        kill $NODE_PID
+    fi
 }
-trap kill_watchers EXIT
 
 main() {
     echo -e "\033[1m\033[34mBuilding API Server\033[0m"
     build_server
     echo
 
+    echo -e "\033[1m\033[34mBuilding lib\033[0m"
+    build_lib
+    echo
+
+    trap kill_watchers EXIT
     echo -e "\033[1m\033[34mWatching Application Source\033[0m"
     watch_app_source
     echo
